26 Jun 2023: added confirm to delete.js

+++ model/delete.js
@@ -5,9 +5,11 @@ export function deleteCard() {
 
   const index = newData.index - 1;
 
-  data.splice(index, 1);
+  if (confirm("Are you sure you want to delete this entry?")) {
+    data.splice(index, 1);
 
-  sessionStorage.setItem("data", JSON.stringify(data));
+    sessionStorage.setItem("data", JSON.stringify(data));
 
-  window.location.replace("#/");
+    window.location.replace("#/");
+  }
 }
